requirements.txt
fastapi<=0.99.1
httpx<=0.25.2
Pillow<=10.1.0
pydantic<=1.10.9
pytest<=7.4.3
python-multipart
uvicorn<=0.22.0